Enhance combo data
This includes adding the core classes to combos (e.g. a spell that
gives bonus abilities to Wizards now lists Wizards). This used to not
be the case, with the assumption that core cards are always in play. However,
adding this means that it can have an impact on the Any slot. Hence, this fixes #1.

Additional improvements to the combo and meta data are included as a result
of an audit, including fixing #6 by adding data for cards that can
destroy cards or that respond to / prevent destruction.