Migrate source code to null safety
After having tried automatic migration, I decided to migrate the code
by hand. There are several assumptions built into the code that are
not provably sound, such as that a card will be built before it is
displayed. Right now, this is required since the overall flow of code
was maintained: a template card is created and 'filled in' as it
passes through the construction process. This is sufficient for now,
although it would be interesting in the future to implement something
more like a factory or builder pattern to avoid the assumptions
elsewhere in the code (e.g. in the UI, which assumes all the pieces
have been put in place.)