Add Barricades Mode setting
This mode reports the Guardians as level VII and filters out a small
set of cards as defined in the barricades blacklist. These cards all
involve either final score (Spices), which is not appropriate for
cooperative play, or making Dungeon phases after Village phases
(Scathian, Open Sesame), which is redundant for Barricades mode.

This fixes #10.