Add localization support
As described in the README file, localization comes in two forms: the
cards database can be localized by inserting translation data, and the
application itself can be localized using the Intl package.

There are not currently any other languages included: this change is
providing the framework whereby new languages can be added.