Introduce preferences API to reduce redundancy in SettingsModel
The new implementation uses source_gen to generate part of the
SettingsModel class.  When the sources are generated via build_runner,
it will go through the SettingsModel and generate an extension method
that provides in an array the fields that are Preference implementors,
along with appropriate accessors and mutators. See the README
for more information about VSCode integration with build_runner.

There is one simple integration test that just ensures that at least
one preference is read from the SettingsModel.