Add working proof-of-concept preflist generator using source_gen
The only preference currently handled this way is showMemo.

When the sources are generated via build_runner (ctrl-shift-b in
vscode), it will go through the SettingsModel and generate an
extension method that provides in an array the fields that are of type
BoolPreference.

The unit test ensures that at least one element is generated.