Remove separate test dependency
With an upgrade to flutter, there was an error thrown by using
both test and flutter_test dependencies. Removing this line seemed
to make it work, although this coincides with some other changes
that, honestly, I am a bit confused about. I'm pushing this out
to GitHub in hopes that this fixes the broken build.