Update dependencies and README
Something in the dependency upgrade regarding code generation caused
the build to fail. Updating the build.yaml file for the preflist
generator (adding a simple ".g") got it working again.

All the automatic fixes from `dart fix` have also been applied.

The settings-related test case has been updated to work properly as
well. Again, this is likely related to changes in dependencies, and it
was an easy fix.